feat: 角色框架 + 难度系统 + DPS测试 + 引爆重写 + 代码优化
架构：
- ICharacterPassive/IGunState/ICharacterConfig/IEvolutionHandler 接口
- CharacterFactory 角色工厂
- DetonateSystem 泛型化（MagePassive→ICharacterPassive）
- GameReferences.CharacterPassive 泛型属性

难度系统：
- DifficultyConfig 10阶参数
- DifficultyManager 选择/倍率/解锁/存档
- DifficultySelectUI 难度选择界面
- WaveAffixSystem 6种波次词缀
- 每波敌人+1护甲 + 混合护甲公式

引爆重写：
- DetonateWaveEffect 红色冲击波扩展
- 时停机制（Time.timeScale=0）
- 26个可配置参数在 DotEffectConfig

DPS测试：
- 主菜单G键进入
- TrainingDummy 20万HP木桩
- DpsTracker 实时DPS面板

伤害float化：
- Damageable.TakeDamage 改为float
- DamagePopup F2格式显示
- StatusEffectManager.Detonate 返回float

强化精简：
- 删除15种强化，新增10种一般强化
- 修复腐蚀/凋零无效果bug
- 风子弹固定3发0.5s冷却

代码优化：
- PhysicsHelper零分配（19处）
- GUIStyle缓存（12文件60处）
- DotSpreadHelper传播去重
- MenuButtonFactory提取
- MageUpgradeApplier策略字典
- 目录重组（UI 5子目录, Core 4子目录）

删除：DotFusionSystem, DotStatusIndicator, DotStatusIconManager

+++ Assets/Scripts/Combat/StatusEffects/StatusEffectData.cs
@@ -44,7 +44,7 @@ public class StatusEffect
 /// </summary>
 public struct DetonateResult
 {
-    public int totalDamage;
+    public float totalDamage;
     public int poisonStacks, burnStacks, bleedStacks, frostStacks;
     public bool hadBurn, hadFrost, hadPoison;
 }
